move BYODModal into the BYODConnectors directory
